projects/cle/src/requirements-dev.txt
python-fasthtml==0.6.9